migrating assesment result functions to seperate go files
