+++ service/orchestrator/orchestrator.go
@@ -39,8 +39,6 @@ import (
 	"clouditor.io/clouditor/persistence"
 	"clouditor.io/clouditor/persistence/inmemory"
 	"clouditor.io/clouditor/service"
-	"golang.org/x/exp/maps"
-	"golang.org/x/exp/slices"
 
 	"github.com/sirupsen/logrus"
 	"google.golang.org/grpc/codes"
@@ -274,43 +272,6 @@ func (s *Service) StoreAssessmentResults(stream orchestrator.Orchestrator_StoreA
 			return status.Errorf(codes.Unknown, "%v", newError.Error())
 		}
 	}
-}
-
-// ListAssessmentResults is a method implementation of the orchestrator interface
-func (svc *Service) ListAssessmentResults(ctx context.Context, req *assessment.ListAssessmentResultsRequest) (res *assessment.ListAssessmentResultsResponse, err error) {
-	var values = maps.Values(svc.results)
-	var filtered []*assessment.AssessmentResult
-	var allowed []string
-	var all bool
-
-	// Check, if the current filter is valid according to the authorization strategy. Omitting the cloud service ID is
-	// only allowed, if one can access all the cloud services. Furthermore, the content of the filtered cloud service ID
-	// must be in the list of allowed cloud service IDs.
-	all, allowed = svc.authz.AllowedCloudServices(ctx)
-	if (req.FilteredCloudServiceId == "" && !all) ||
-		(req.FilteredCloudServiceId != "" && !slices.Contains(allowed, req.FilteredCloudServiceId)) {
-		return nil, service.ErrPermissionDenied
-	}
-
-	for _, v := range values {
-		if req.FilteredCloudServiceId != "" && v.CloudServiceId != req.FilteredCloudServiceId {
-			continue
-		}
-
-		filtered = append(filtered, v)
-	}
-
-	res = new(assessment.ListAssessmentResultsResponse)
-
-	// Paginate the results according to the request
-	res.Results, res.NextPageToken, err = service.PaginateSlice(req, filtered, func(a *assessment.AssessmentResult, b *assessment.AssessmentResult) bool {
-		return a.Timestamp.AsTime().After(b.Timestamp.AsTime())
-	}, service.DefaultPaginationOpts)
-	if err != nil {
-		return nil, status.Errorf(codes.Internal, "could not paginate results: %v", err)
-	}
-
-	return
 }
 
 func (s *Service) RegisterAssessmentResultHook(hook func(result *assessment.AssessmentResult, err error)) {
